Implement new command "shuffle" Origionally there is no operation to shuffle the queue. By using "Fisher-Yates shuffle" algorithm. A new command "shuffle" is implmented.
In "shuffle.c", the fuctionality is implemented by using "Fisher-Yates
shuffle" algorithm. Randomly choosing one node in the queue as
seleted node. And choosing the last node as target node. Through
swapping these two nodes until all node had been choosen. It will
create a shuffled queue.

To determine the performance of it. In "shuffle_driver.py", 10000000
times testing is operated to see the histogram of the shuffled queue.
Also the chi-squared test statistic is calculated to evaluate the
performance.