Implement the arithmetic of fixed point operation   In Linux Kernel, the calculations of floating points should be carefully used. Since there are machines without FPU. Moreover, in general, fixed point operations could be faster than their floating point counterparts.
  Thus, a set of fixed-point arithmetic is introduced here to examine
the basic calculation of fixed-point.

  A data structure called "fixed_point_t" is defined. Which the lower
10 bits are used to represent decimal part of the fixed points. A
set of macros such as "int2fix" are used to transform between
"int" and "fixed_point_t" by shiftting. "extract_frac" and "frac2int"
can be used to extract the decimal part of the value down to three
digits. In "fixed_div", the numerator part is first rigtht-shiffted
before the calculation to prevent the rounding effect caused by
interger operation.